audio-rag-agent-deploy/requirements.txt
python-dotenv>=0.19.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
python-multipart>=0.0.5
sounddevice>=0.4.5
numpy>=1.21.0
scipy>=1.9.0
transformers>=4.30.0
torch>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-groq>=0.1.0
sentence-transformers>=2.5.0
chromadb>=0.4.0
python-docx>=0.8.11
PyPDF2>=3.0.0
soundfile>=0.12.1
webrtcvad>=2.0.10
yt-dlp>=2023.10.0
python-ffmpeg>=2.0.0
ffmpeg-python>=0.2.0
librosa>=0.10.1

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
httpx>=0.24.1
pytest-mock>=3.11.1
aiofiles>=23.1.0
pytest-env>=1.0.1
